Adds DbRequestData, changes RequestObject to JavaBean
Added the DbRequestData object, which will be used to send RequestObjects over the network from client to server and back. To facilitate this, RequestObject has been converted into a JavaBean.
There is one FindBugs error that is still yelling at me, because it doesn't think that RequestObject is serializable, wich it totally is.