Added right-click functionality to Users table in the Users tab.
When a row in the table is right-clicked, a context menu will pop up with
two options, "Send Message" and "View Transactions"

ServerGui.java has been altered to make the tabbed pane that holds all of the tabs
accessible from outside of the ServerGui.java class. This was necessary to
implement the functionality to switch tabs from the
CreateButtonsSearch.java class.

In the CreateButtonsSearch.java class, implemented three listeners, two to
switch tabs and one to select the row that was right-clicked.