Created DatabaseHelper class to access database.
Connecting to the database should be abstract to enable reusability. The
class contains methods to retrieve a set of users, update a specific user,
and delete a specific user from the database, and run a custom statement
on the database.

Also, previous build's failure was fixed by adding final keyword to
messageText in MessagingTab.